Extract common elements
